# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: BehaveDrift API
  version: 1.0.0
  description: |
    **BehaveDrift API** is an open-source AI inference engine that detects
    behavioural pattern drift in dementia residents.

    It accepts longitudinal care observation data and returns:
    - A **drift score** quantifying deviation from each resident's personal baseline
    - **Tiered alerts** (T1–T4) with natural-language explanations
    - **FHIR R4** Condition resources for clinical handover

    ### Authentication

    Two modes are supported:

    **OAuth2 Client Credentials (recommended for production):**
    ```
    POST /v1/auth/token
    grant_type=client_credentials&client_id=...&client_secret=...
    Authorization: Bearer <jwt>
    ```

    **API Key (for self-hosted / development):**
    ```
    X-API-Key: bda_sk_xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
    ```

    ### Error Responses
    All errors use a consistent shape:
    ```json
    {
      "error": "validation_error",
      "message": "Field 'resident_id' is required.",
      "code": 422,
      "details": [],
      "request_id": "req_abc123"
    }
    ```

    ### Versioning
    All endpoints are prefixed with `/v1/`. Breaking changes increment to `/v2/`.

  contact:
    name: Umar Isa
    email: hello@behavedrift.io
    url: https://github.com/umarisa/behavedrift-api

  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OAuth2 Client Credentials JWT (RS256, 60 min TTL)

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key for self-hosted / development use

  # ============================================================
  # SHARED SCHEMAS
  # ============================================================
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, message, code, request_id]
      properties:
        error:
          type: string
          description: Machine-readable error code (snake_case)
          example: validation_error
        message:
          type: string
          description: Human-readable explanation
          example: "Field 'resident_id' is required and cannot be blank."
        code:
          type: integer
          description: HTTP status code
          example: 422
        details:
          type: array
          description: Field-level validation errors (422 only)
          items:
            type: object
            properties:
              field:
                type: string
                example: resident_id
              issue:
                type: string
                example: required
              value:
                nullable: true
                example: null
        request_id:
          type: string
          description: Unique request ID for log correlation
          example: req_4f3a9c21d
        documentation_url:
          type: string
          format: uri
          example: https://docs.behavedrift.io/errors/validation_error

    # ---------- Pagination ----------
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 142
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 20
        has_next:
          type: boolean
          example: true

    # ---------- Resident ----------
...
    MoodSignal:
      type: object
      properties:
        value:
          type: integer
          minimum: 1
          maximum: 5
          description: "1=very low, 3=neutral, 5=very positive"
          example: 2
        scale:
          type: string
          default: "1-5"
        notes:
          type: string
          maxLength: 300
          example: "Noticeably more agitated than usual this afternoon"
    AppetiteSignal:
      type: object
      properties:
        value:
          type: string
          enum: [excellent, good, fair, poor, refused]
          example: poor
        meal:
          type: string
          enum: [breakfast, lunch, dinner, snack]
          example: lunch
        fluid_intake_ml:
          type: integer
          minimum: 0
          example: 120
    SleepSignal:
      type: object
      properties:
        value:
          type: string
          enum: [good, fair, disturbed, very_disturbed, unknown]
          example: disturbed
        night_wakings:
          type: integer
          minimum: 0
          example: 4
        hours_slept:
          type: number
          minimum: 0
          maximum: 24
          example: 4.5
    SocialEngagementSignal:
      type: object
      properties:
        value:
          type: string
          enum: [engaged, moderate, withdrawn, isolated]
          example: withdrawn
        activity_participated:
          type: boolean
          example: false
        activity_name:
          type: string
          example: morning_exercise
    PainIndicatorSignal:
      type: object
      properties:
        facial_grimacing:
          type: boolean
          example: false
        guarding:
          type: boolean
          example: false
        vocalisation:
          type: boolean
          example: true
        restlessness:
          type: boolean
          example: true
        verbal_report:
          type: boolean
          example: false
        painad_score:
          type: integer
          minimum: 0
          maximum: 10
          description: PAINAD scale score (optional, for clinical use)
          example: 4
    MobilitySignal:
      type: object
      properties:
        value:
          type: string
          enum: [independent, supervised, assisted, dependent, bedbound]
          example: assisted
        baseline_comparison:
          type: string
          enum: [better_than_usual, same_as_usual, worse_than_usual, unknown]
          example: worse_than_usual
    AgitationSignal:
      type: object
      properties:
        value:
          type: string
          enum: [calm, mild, moderate, severe]
          example: moderate
        type:
          type: string
          enum: [physical, verbal, wandering, sundowning, other]
          example: sundowning

    # ---------- Observation ----------
    ObservationSignals:
      type: object
      description: At least one signal must be provided
      minProperties: 1
      properties:
        mood:
          $ref: '#/components/schemas/MoodSignal'
        appetite:
          $ref: '#/components/schemas/AppetiteSignal'
        sleep_quality:
          $ref: '#/components/schemas/SleepSignal'
        social_engagement:
          $ref: '#/components/schemas/SocialEngagementSignal'
        pain_indicators:
          $ref: '#/components/schemas/PainIndicatorSignal'
        mobility:
          $ref: '#/components/schemas/MobilitySignal'
        agitation:
          $ref: '#/components/schemas/AgitationSignal'
    ObservationContext:
      type: object
      properties:
        location:
          type: string
          example: bedroom
        visitor_present:
          type: boolean
          example: false
        medication_administered:
          type: boolean
          example: true
        medication_notes:
          type: string
          maxLength: 200
          example: Routine morning meds given
        staff_id:
          type: string
          description: Anonymised staff reference (not stored in logs)
          example: staff_42b
    ObservationCreate:
      type: object
      required: [resident_id, observed_at, signals]
      properties:
        resident_id:
          type: string
          example: res_7f3a9c21
        observed_at:
          type: string
          format: date-time
          example: "2026-03-14T14:30:00Z"
        observer_id:
          type: string
          description: Anonymised staff reference
          example: staff_42b
        signals:
          $ref: '#/components/schemas/ObservationSignals'
        context:
          $ref: '#/components/schemas/ObservationContext'
    DriftEvaluationResult:
      type: object
      properties:
        triggered:
          type: boolean
          example: true
        drift_score:
          type: number
          minimum: 0
          maximum: 1
          example: 0.67
        baseline_status:
          type: string
          enum: [initialising, active, reset_pending, suspended]
          example: active
        signals_flagged:
          type: array
          items:
            type: string
          example: [mood, appetite, social_engagement]
        alert_generated:
          type: object
          nullable: true
          properties:
            alert_id:
              type: string
              example: alr_3c8e1b90
            tier:
              type: string
              enum: [T1, T2, T3, T4]
              example: T2
            tier_label:
              type: string
              example: Concern
        message:
          type: string
          description: Human-readable status note (e.g. when baseline is initialising)
          example: null
    ObservationResponse:
      type: object
      properties:
        observation_id:
          type: string
          example: obs_a9d2f1c3
        resident_id:
          type: string
          example: res_7f3a9c21
        processed_at:
          type: string
          format: date-time
          example: "2026-03-14T14:30:04Z"
        drift_evaluation:
          $ref: '#/components/schemas/DriftEvaluationResult'
        status:
          type: string
          enum: [processed, queued, failed]
          example: processed
    ObservationBatchRequest:
      type: object
      required: [observations]
      properties:
        observations:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: '#/components/schemas/ObservationCreate'
    ObservationBatchResponse:
      type: object
      properties:
        submitted: { type: integer, example: 10 }
        processed: { type: integer, example: 9 }
        failed: { type: integer, example: 1 }
        results:
          type: array
          items:
            $ref: '#/components/schemas/ObservationResponse'
        errors:
          type: array
          items:
            type: object
            properties:
              index: { type: integer }
              error: { type: string }
              message: { type: string }

    # ---------- Alert ----------
...
  responses:
    Unauthorized:
      description: Authentication required or token expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: authentication_required
            message: A valid Bearer token or X-API-Key header is required.
            code: 401
            request_id: req_abc123
...
    NotFound:
      description: Resource not found (or belongs to another tenant)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: not_found
            message: The requested resource does not exist.
            code: 404
            request_id: req_abc123
    UnprocessableEntity:
      description: Request body failed schema validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: validation_error
            message: "Field 'resident_id' is required."
            code: 422
            details:
              - field: resident_id
                issue: required
                value: null
            request_id: req_abc123
    RateLimited:
      description: Too many requests
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: rate_limit_exceeded
            message: You have exceeded 300 requests per minute. Retry after 30 seconds.
            code: 429
            request_id: req_abc123
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: internal_error
            message: An unexpected error occurred. The request_id has been logged.
            code: 500
            request_id: req_abc123

# ============================================================
# SECURITY (applied globally by default)
# ============================================================
security:
  - BearerAuth: []
  - ApiKeyAuth: []

# ============================================================
# PATHS
# ============================================================
paths:
...
  /v1/observations:
    post:
      tags: [Observations]
      summary: Submit a single observation
      description: |
        Submit one behavioural observation for a resident.
        The drift engine evaluates it immediately and returns a score.
        If the drift score exceeds the configured threshold, an alert is generated.
      operationId: createObservation
      security:
        - BearerAuth: [behavedrift:write]
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ObservationCreate'
      responses:
        '201':
          description: Observation processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObservationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: resident_id not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/observations/batch:
    post:
      tags: [Observations]
      summary: Submit a batch of observations
      description: |
        Submit up to 100 observations in a single request.
        Useful for bulk backfill or syncing from another system.
        Each observation is processed independently — partial success is possible.
      operationId: createObservationBatch
      security:
        - BearerAuth: [behavedrift:write]
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ObservationBatchRequest'
      responses:
        '200':
          description: Batch processed (check `errors` array for individual failures)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObservationBatchResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /v1/observations/fhir:
    post:
      tags: [Observations]
      summary: Submit a FHIR R4 Observation
      description: |
        Accept a FHIR R4 Observation resource directly.
        BehaveDrift maps it to the internal signal schema automatically.
        Supported SNOMED codes map to BehaveDrift signals — see docs/fhir-mapping.md.
      operationId: createFhirObservation
      security:
        - BearerAuth: [behavedrift:write]
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/fhir+json:
            schema:
              type: object
              description: A valid FHIR R4 Observation resource
              required: [resourceType, status, code, subject, effective]
              properties:
                resourceType:
                  type: string
                  enum: [Observation]
                status:
                  type: string
                  enum: [final, amended, preliminary]
                code:
                  type: object
                  description: SNOMED CT coded observation type
                subject:
                  type: object
                  description: Reference to Patient (resident_id)
                effective:
                  type: string
                  format: date-time
      responses:
        '201':
          description: FHIR Observation processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObservationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          description: FHIR resource invalid or SNOMED code unmapped
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/observations/{resident_id}:
    get:
      tags: [Observations]
      summary: Get observation history for a resident
      operationId: getObservations
      security:
        - BearerAuth: [behavedrift:read]
        - ApiKeyAuth: []
      parameters:
        - name: resident_id
          in: path
          required: true
          schema:
            type: string
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
        - name: signal
          in: query
          description: Filter to a specific signal type
          schema:
            type: string
            enum: [mood, appetite, sleep_quality, social_engagement, pain_indicators, mobility, agitation]
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: page_size
          in: query
          schema: { type: integer, default: 50, maximum: 200 }
      responses:
        '200':
          description: Paginated observation history
          content:
            application/json:
              schema:
                type: object
                properties:
                  resident_id: { type: string }
                  observations:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/ObservationCreate'
                        - type: object
                          properties:
                            observation_id: { type: string }
                            processed_at: { type: string, format: date-time }
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---- Alerts ----
...